Removed VBox, renamed stylable vbox to vbox, made changes to match with HBox PR
