--- packages/wix-ui-core/src/components/StylableVBox/index.ts
@@ -1,1 +1,0 @@
-export {VBox, VBoxProps} from './VBox';
