+++ packages/wix-ui-core/src/components/VBox/VBox.e2e.ts
@@ -1,7 +1,7 @@
 import * as eyes from 'eyes.it';
 import {browser} from 'protractor';
 import {getStoryUrl, waitForVisibilityOf} from 'wix-ui-test-utils/protractor';
-import {vboxTestkitFactory} from '../../testkit/protractor';
+import {vBoxTestkitFactory} from '../../testkit/protractor';
 
 describe('VBox', () => {
   const storyUrl = getStoryUrl('Components', 'VBox');
@@ -9,7 +9,7 @@ describe('VBox', () => {
   beforeEach(() => browser.get(storyUrl));
   eyes.it('should display correct content', () => {
     const dataHook = 'storybook-vbox';
-    const driver = vboxTestkitFactory({dataHook});
+    const driver = vBoxTestkitFactory({dataHook});
 
     return waitForVisibilityOf(driver.element(), 'Cannot find VBox');
   });
