+++ packages/wix-ui-core/src/components/VBox/VBox.tsx
@@ -1,15 +1,34 @@
 import * as React from 'react';
-import {createHOC} from '../../createHOC';
+import {number, any, oneOf, arrayOf} from 'prop-types';
+import style from './VBox.st.css';
+import {addSpacing} from './utils';
 
 export interface VBoxProps {
-  children: any;
-  classes?: {
-    vRoot: string
-  };
+  children?: React.ReactNode;
+  horizontalAlignment?: Alignment;
+  spacing?: number;
 }
 
-const VBox: React.SFC<VBoxProps> = ({children, classes}) => (
-  <div className={classes.vRoot}>{children}</div>
-);
+export type Alignment = 'left' | 'center' | 'right';
 
-export default createHOC(VBox);
+const defaultProps: VBoxProps = {
+  children: null,
+  horizontalAlignment: 'left',
+  spacing: 0
+};
+
+/**
+ * VBox
+ */
+export const VBox: React.SFC<VBoxProps> = props => {
+  const {children, horizontalAlignment, spacing} = props;
+  return <div {...style('root', {horizontalAlignment}, props)}>{addSpacing(children, spacing)}</div>;
+};
+
+VBox.defaultProps = defaultProps;
+
+VBox.propTypes = {
+  children: any,
+  horizontalAlignment: oneOf(['left', 'center', 'right']),
+  spacing: number
+};
