+++ packages/wix-ui-core/src/components/VBox/index.ts
@@ -1,9 +1,1 @@
-import * as React from 'react';
-import VBoxComponent, {VBoxProps} from './VBox';
-import {styles} from './styles';
-import {withClasses} from 'wix-ui-jss';
-import {ThemedComponentProps} from 'wix-ui-jss';
-import {WixComponentProps} from '../../createHOC';
-
-export {VBoxProps};
-export const VBox = withClasses(VBoxComponent, styles) as React.ComponentClass<VBoxProps & ThemedComponentProps & WixComponentProps>;
+export {VBox, VBoxProps} from './VBox';
