--- packages/wix-ui-core/src/components/VBox/theme.ts
@@ -1,9 +1,0 @@
-export type VBoxTheme = {
- spacing?: string
- horizontalAlignment?: 'left' | 'center' | 'right'
-};
-
-export const core: VBoxTheme = {
-  spacing: '20px',
-  horizontalAlignment: 'left'
-};
